requirements.txt
anyio==4.12.1
certifi==2026.1.4
charset-normalizer==3.4.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
markdown-it-py==4.0.0
mdurl==0.1.2
Pygments==2.19.2
requests==2.32.5
rich==14.2.0
typing_extensions==4.15.0
urllib3==2.6.3
python-dotenv==1.0.1
